% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrowhead_gm12878_5kb.R
\docType{data}
\name{arrowhead_gm12878_5kb}
\alias{arrowhead_gm12878_5kb}
\title{Domain data from ARROWHEAD TAD-caller for GM12878 at 5 kb}
\format{
An object of class \code{data.frame} with 8409 rows and 3 columns.
}
\source{
Data from Rao SS, Huntley MH, Durand NC, Stamenova EK et al.
A 3D map of the human genome at kilobase resolution reveals principles
of chromatin looping. Cell 2014 Dec 18;159(7):1665-80. PMID: 25497547.
Available at
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE63525}
}
\usage{
arrowhead_gm12878_5kb
}
\value{
A data.frame
}
\description{
A data frame with 3 columns and 8409 rows
\describe{
\item{V1}{The chromosome number}
\item{V2}{The start coordinate of the TAD}
\item{V3}{The end coordinate of the TAD}
}
}
\keyword{datasets}
