% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_func.R
\name{binary_func}
\alias{binary_func}
\title{Helper function used to create binary overlap type feature space}
\usage{
binary_func(binned_data_gr, annot_data_gr)
}
\arguments{
\item{binned_data_gr}{A GRanges object}

\item{annot_data_gr}{A Granges object}
}
\value{
An indicator vector for whether or not an overlap occured
}
\description{
Helper function used to create binary overlap type feature space
}
