% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_func.R
\name{count_func}
\alias{count_func}
\title{Helper function used to create count overlap type feature space}
\usage{
count_func(binned_data_gr, annot_data_gr)
}
\arguments{
\item{binned_data_gr}{A GRanges object}

\item{annot_data_gr}{A Granges object}
}
\value{
A vector of counts enumerating the number of overlaps
}
\description{
Helper function used to create count overlap type feature space
}
