% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTADdata.R
\name{createTADdata}
\alias{createTADdata}
\title{Function to create a data matrix used for building a predictive model to
classify boundary regions from functional genomic elements}
\usage{
createTADdata(
  bounds.GR,
  resolution,
  genomicElements.GR,
  featureType = "distance",
  resampling,
  trainCHR,
  predictCHR = NULL,
  seed = 123
)
}
\arguments{
\item{bounds.GR}{a GRanges object with chromosomal coordinates of TAD
boundaries used to identify positive cases (can be obtained using
\code{\link{extractBoundaries}})}

\item{resolution}{Numeric, the width to bin the genome at, should match the
resolution that TADs were called at}

\item{genomicElements.GR}{a GRangesList object containing GRanges objects
for each ChIP-seq data to leverage in the random forest model (can be
obtained using the \code{\link{bedToGRangesList}})}

\item{featureType}{Character, controls how the feature space is constructed
(one of either "binary", "oc", "op", or "distance" (log2-transformed);
default is "distance")}

\item{resampling}{Character, controls if and how the data should be
resampled to create balanced classes of boundary vs nonboundary regions (one
of either "none"-no re-sampling, "ros"-Random Over-Sampling, "rus"-Random
Under-Sampling, or "smote"-Synthetic Minority Over-sampling TEchnique)}

\item{trainCHR}{Character vector of chromosomes to use to build the binned
data matrix for training}

\item{predictCHR}{Character vector of chromosomes to use to build the binned
data matrix for testing. Default in NULL, indicating no test data is created.
 If trainCHR=predictCHR then a 7:3 split is created}

\item{seed}{Numeric for reproducibility of resampling}
}
\value{
A list object containing two data.frames: 1) the training data, 2)
the test data (only if predictCHR is not NULL, otherwise it is NA)
}
\description{
Function to create a data matrix used for building a predictive model to
classify boundary regions from functional genomic elements
}
\examples{
\dontrun{
#Create training data for CHR21 and testing data for CHR22 with
#5 kb binning, oc-type predictors from 26 different transcription factor
#binding sites from the GM12878 cell line, and random under-sampling

#Read in ARROWHEAD-called TADs at 5kb
data(arrowhead_gm12878_5kb)

#Extract unique boundaries
bounds.GR <- extractBoundaries(domains.mat=arrowhead_gm12878_5kb,
                               preprocess=FALSE,
                               CHR=c("CHR21","CHR22"),
                               resolution=5000)

#Read in GRangesList of 26 TFBS
data(tfbsList)

tadData <- createTADdata(bounds.GR=bounds.GR,
                         resolution=5000,
                         genomicElements.GR=tfbsList,
                         featureType="oc",
                         resampling="rus",
                         trainCHR="CHR21",
                         predictCHR="CHR22",
                         seed=123)
}
}
