% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_func.R
\name{distance_func}
\alias{distance_func}
\title{Helper function used to create (log2) distance type feature space}
\usage{
distance_func(binned_data_center_gr, annot_data_center_gr)
}
\arguments{
\item{binned_data_center_gr}{A GRanges object of width 1}

\item{annot_data_center_gr}{A GRanges object of width 1}
}
\value{
A vector of log2 distances to the nearest overlap
}
\description{
Helper function used to create (log2) distance type feature space
}
