% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_func.R
\name{percent_func}
\alias{percent_func}
\title{Helper function used to create percent overlap type feature space}
\usage{
percent_func(binned_data_gr, annot_data_gr)
}
\arguments{
\item{binned_data_gr}{A GRanges object}

\item{annot_data_gr}{A Granges object}
}
\value{
A vector of proportions indicating the percentage of overlap
}
\description{
Helper function used to create percent overlap type feature space
}
