% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_func.R
\name{signal_func}
\alias{signal_func}
\title{Helper function used to create signal type feature space}
\usage{
signal_func(binned_data_gr, annot_data_gr)
}
\arguments{
\item{binned_data_gr}{A GRanges object}

\item{annot_data_gr}{A GRanges object}
}
\value{
A vector of intensities indicating the signal strength within each
overlap
}
\description{
Helper function used to create signal type feature space
}
