% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBoundaries.R
\name{extractBoundaries}
\alias{extractBoundaries}
\title{Function to extract using boundaries from domain data}
\usage{
extractBoundaries(domains.mat, preprocess = FALSE, CHR, resolution)
}
\arguments{
\item{domains.mat}{either a \code{matrix} or \code{data.frame} with at least
3 coluns. First column is either of class "numeric" or "integer". The second
and third columns are the start and end coordinates of domains, respectively.}

\item{preprocess}{logical, indicating whether or not domains exceeding 2mb
in width or smaller than 2*(the specified resolution) should be filtered out
 (default is FALSE, given that prediction is boundary-specific)}

\item{CHR}{character, specifying which chromosome(s) to extract domain
boundaries on (ex: "CHR22")}

\item{resolution}{numeric, the Hi-C data resolution that domains were called
at}
}
\value{
A \code{GRanges} object
}
\description{
Function to extract using boundaries from domain data
}
\examples{
#Read in domain data from ARROWHEAD at 5 kb for GM12878
data("arrowhead_gm12878_5kb")
#Extract unique boundaries for CHRs 1-8 and 10-22
bounds.GR <- extractBoundaries(domains.mat=arrowhead_gm12878_5kb,
                               preprocess=FALSE,
                               CHR=paste0("CHR",c(1:8,10:22)),
                               resolution=5000)
}
