% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juicer_func.R
\name{juicer_func}
\alias{juicer_func}
\title{Helper function for transforming a GRanges object into matrix form to be
saved as .txt or .BED file and imported into juicer}
\usage{
juicer_func(grdat)
}
\arguments{
\item{grdat}{A GRanges object}
}
\value{
A matrix that can be saved as a BED file to import into juicer
}
\description{
Helper function for transforming a GRanges object into matrix form to be
saved as .txt or .BED file and imported into juicer
}
